BaseScoutingActivity.relativeTime: Use Double to avoid conversion in fields

+++ app/src/main/java/ca/warp7/android/scouting/BaseScoutingActivity.kt
@@ -18,7 +18,7 @@ interface BaseScoutingActivity {
     /**
      * Relative time
      */
-    val relativeTime: Int
+    val relativeTime: Double
 
     /**
      * The time Handler of the activity
